+++ Changelog.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased: 5.0.0-rc.100]
+ - FIO-9360: validate current page only on wizard change
+
 ## 5.0.0-rc.99
 ### Changed
  - Updated @formio/core@2.3.0-rc.22